Fixing bugs

+++ src/common/common.h
@@ -57,7 +57,7 @@ inline static uint8_t bit8_clear(uint8_t value, int start, int bits) { return va
 
 uint64_t bit_gen(const char *s, int bits); // Generate 64 bit string using "1" and "0", MSB first
 inline static uint8_t bit8_gen(const char *s) { return (uint8_t)bit_gen(s, 8); }
-inline static uint8_t bit64_gen(const char *s) { return (uint8_t)bit_gen(s, 64); }
+inline static uint64_t bit64_gen(const char *s) { return bit_gen(s, 64); }
 
 // Copies bit range in one byte to another byte
 // starts are inclusive; ends are non-inclusive
